\name{mcode-package}
\alias{mcode-package}
\title{Functions to merge and recode across multiple variables}
\description{}
\details{
\tabular{ll}{
Package: \tab mcode\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-10\cr
License: \tab GPL-2\cr
}
}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}